core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.01.00"
description = "Pipeline Core Module"
authors = ["Meddy Boukhedouma <meddy.boukhedouma@gmail.com>"]

[tool.poetry.dependencies]
python = ">=2.7,<3.0.0 || >=3.7.0,<3.10"
jeanpaulstart = {git = "https://github.com/Arubinu/jeanpaulstart.git", rev = "master"}
jeanpaulstartui = {git = "https://github.com/Pixels-Coder/jeanpaulstartui.git", rev = "feature/python3_qt5"}
"Qt5.py" = "^0.1.0"
PySide2 = {version = "^5.15.2", optional = true }
PyYAML = "<6.0"
Lucidity = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=44.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
ui = ["PySide2"]

[tool.poetry.scripts]
flow = "core.__main__:main"